added parse function, jquery, parse content script
parse.js just has a line of code to hide the entire site right now. Manifest has been changed to include the content scripts. The "site-container" is the parent class of all the elements we need to find in the DOM. I will start their and move down the descendants until I hit the "wprm-recipe-container" class which is what we need.